Use type and properties objects for arrays, tweak required field syntax
Many arrays defined in GBFS schemas define item properties directly in
the containing object. In this commit, we provide a schema for
these array items to be validated against. Because in most cases array
items are objects, this amounted to adding "type": "object" and moving
the properties that were defined under a "properties" key.

Additionally, in some cases, array properties would be marked required
by setting "required": "true" below the "items" property. We also now
use the more common syntax of requiring fields by setting "required" to
a list of strings.